TFM/cloud-functions/course-pipeline-job/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.34.0
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.100.0